+++ lib/elixir/lib/binary/dict.ex
@@ -1,9 +1,8 @@
-defmodule Orddict do
+defmodule Binary.Dict do
   @moduledoc """
-  This module implements a dictionary based that stores items
-  as a list of tuples. It is a simple wrapper around
-  [Erlang's orddict module](http://www.erlang.org/doc/man/orddict.html)
-  and exposed via the `Dict` module.
+  This module implements a dictionary that forces the keys to be
+  converted to binaries on insertion. Currently it is implemented
+  using an `Orddict`, but this may change in the future.
 
   Check the `Dict` module for examples and documentation.
   """
@@ -12,7 +11,7 @@ defmodule Orddict do
 
   defmacrop dict(data) do
     quote do
-      { Orddict, unquote(data) }
+      { Binary.Dict, unquote(data) }
     end
   end
 
@@ -29,22 +28,22 @@ defmodule Orddict do
   end
 
   def has_key?(dict(data), key) do
-    :orddict.is_key key, data
+    :orddict.is_key to_binary(key), data
   end
 
   def get(dict(data), key, default // nil) do
-    case :orddict.find(key, data) do
+    case :orddict.find(to_binary(key), data) do
       {:ok, value} -> value
       :error       -> default
     end
   end
 
   def put(dict(data), key, value) do
-    dict(:orddict.store key, value, data)
+    dict(:orddict.store to_binary(key), value, data)
   end
 
   def delete(dict(data), key) do
-    dict(:orddict.erase key, data)
+    dict(:orddict.erase to_binary(key), data)
   end
 
   def merge(dict(d1), dict(d2), fun) do
@@ -56,11 +55,11 @@ defmodule Orddict do
   end
 
   def update(dict(data), key, fun) do
-    dict(:orddict.update key, fun, data)
+    dict(:orddict.update to_binary(key), fun, data)
   end
 
   def update(dict(data), key, initial, fun) do
-    dict(:orddict.update key, fun, initial, data)
+    dict(:orddict.update to_binary(key), fun, initial, data)
   end
 
   def empty(_) do
@@ -72,16 +71,16 @@ defmodule Orddict do
   end
 end
 
-defimpl Enum.Iterator, for: Orddict do
-  def iterator({ Orddict, data }), do: data
-  def count({ Orddict, data }),    do: length(data)
+defimpl Enum.Iterator, for: Binary.Dict do
+  def iterator({ Binary.Dict, data }), do: data
+  def count({ Binary.Dict, data }),    do: length(data)
 end
 
-defimpl Enum.OrdIterator, for: Orddict do
-  def iterator({ Orddict, data }), do: data
+defimpl Enum.OrdIterator, for: Binary.Dict do
+  def iterator({ Binary.Dict, data }), do: data
   def to_list({ h, next }, _),     do: [h|next]
 end
 
-defimpl Access, for: Orddict do
-  def access(dict, key), do: Orddict.get(dict, key, nil)
+defimpl Access, for: Binary.Dict do
+  def access(dict, key), do: Binary.Dict.get(dict, key, nil)
 end